+++ src/components/ContextControlsItem.vue
@@ -3,9 +3,12 @@ import { showBuffer } from '@src/infrastructure/prun-ui/buffers';
 
 const props = defineProps<{ cmd: string; label?: string }>();
 
-const index = props.cmd.lastIndexOf(' ');
-const [command, argument] =
-  index > -1 ? [props.cmd.substring(0, index), props.cmd.substring(index + 1)] : [props.cmd, ''];
+const commandAndArg = computed(() => {
+  const words = props.cmd.split(' ');
+  return words[0] === 'XIT'
+    ? [words[0] + ' ' + words[1], words.slice(2).join(' ') ?? '']
+    : [words[0], words.slice(1).join(' ') ?? ''];
+});
 
 const itemClasses = [C.ContextControls.item, C.fonts.fontRegular, C.type.typeSmall];
 </script>
@@ -14,8 +17,8 @@ const itemClasses = [C.ContextControls.item, C.fonts.fontRegular, C.type.typeSma
   <!-- The node structure is fully replicated from PrUn, don't mind unnecessary nodes. -->
   <div :class="itemClasses" @click="() => showBuffer(cmd)">
     <span>
-      <span :class="C.ContextControls.cmd">{{ command }}</span>
-      {{ argument }}
+      <span :class="C.ContextControls.cmd">{{ commandAndArg[0] }}</span>
+      {{ commandAndArg[1] }}
     </span>
     <span v-if="label" :class="C.ContextControls.label">: {{ label }}</span>
   </div>